refactor AttributeSelectorPopup code, add some tests
The original AttributeSelectorPopup code was almost 1:1 copied to GuidedDecisionTableViewImpl. This removes the duplication
- the AttributeSelectorPopup is now an abstract parent of 2 separate implementations. Shared code (most logic) stayed there.
This also brings the basic metadata input validation that was present in GRE to GDT.
There are now also a few small tests checking that validation and that duplicates are not allowed (part of moving selenium tests
to community).