[GUVNOR-2880] remove unused field causing issues on Tomcat
The biggest issue with the User field was that it was @RequestScoped.
The ScenarioRunnerService is being used as part of AsyncJobs (e.g.
when the project/maven/test REST endpoint is triggered). On Tomcat
this does not work, becuase the request thread is different from the
processing thread and thus the request scoped variables are not
available. For some reason this works on WildFly 10 / EAP7
(most likely because we use the async EJBs there, instead of the
thred pool).