Increase the TableWizard coverage (#739)
This PR focus on a case when the TableWizard informs the user about no available pattenrs which could be used in the table actions. In other words, user can not change a field of a pattern because there is no pattern defined yet.

This PR is part of migration form internal selenium test suite into community. See BAQE-314.