DROOLS-4131: Allow any characters for list operators
This PR fixes the issue in guided rule templates. User was not able to put multiple numeric values separated by comma as the numeric text box was forcing just numbers. This PR change the behavior in the way allowing all characters if the "in" or "not in" operator is selected in the guided rule template.